Add support for fallocate hole punching
It appears ALICE does not track what parts of sparse files are present
or not, because fallocate with mode = 0 does not result in any metadata
about sparseness being tracked, instead it just extends the file's size,
if applicable, and writes zeroes to the range in question. From the
standpoint of durability, presumably it doesn't matter whether the file
system truly marks one block as not present or writes zeroes to it, as
the user will read zeroes from the area either way in the happy case,
and if the operation to zero the block was not persisted, this would
appear the same as if the real operation to mark the block as not
present was not persisted. Thus, falloc(FALLOC_FL_PUNCH_HOLE |
FALLOC_FL_KEEP_SIZE, ...) is implemented here with one write micro
operation across the requested range.